Initial port
This port is based on the shoreline pixelflut server. I've decided on this
port for its efficient parsing routines.
However, a number of properties will not be ideal for the ESP:

 - heavy multithreading will waste performance on dual core system
 - no care taken to produce only efficient 32bit
 - designed for systems with loads of memory available

I may or may not address those issues in the future. Just want to leave
them here so they are at least remembered.